requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
google-cloud-storage==2.13.0
google-generativeai==0.3.2

# Report generation
markdown==3.5.1
pdfkit==1.0.0

# Data handling
pandas==2.1.2

# HTTP client
requests==2.31.0

# Type hinting
typing-extensions==4.8.0